+++ docs/history.md
@@ -5,6 +5,7 @@
 
 ### _v0.2.4_ - October 15th, 2023 (Unreleased)
 
+- A red icon will now show if you are playing on hardcore mode.
 - The format of stats in the save file has changed. If you want to update an old save file to the new format, put the contents of "StatsTracker.persistent" into "StatsTracker.persistent.stats".
 
 ### _v0.2.3_ - October 15th, 2023